docker/requirements.txt
boto3==1.9.127
botocore==1.12.127
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
facebook-sdk==3.1.0
idna==2.8
jmespath==0.9.4
python-dateutil==2.8.0
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
